Refactor: split disasm.rs and stack_trace.rs
